renamed 3 view folders,changed var name to FilmsList
