+++ src/view/button-show-more.js
