+++ src/view/film-list.js
@@ -1,4 +1,4 @@
-export const createFilmsContent = () => {
+export const createFilmsList = () => {
   return (
     `<section class="films">
         <section class="films-list">
