+++ src/view/movies-amount.js
